==> Part2. Step1.1 Training of deep learning models.ipynb <==
# What
分类任务，支持两种模式
1. Folder模式，需要输入`train`, `valid`两个测试集对应的目录。`labels.txt`，需要训练的label，里面每个类别一行。
2. List模式，需要输入`train`, `valid`两个测试集对应的训练文件，每行一个样本。`labels.txt`是可选参数，里面每个类别一行。`data_pattern`一个通用的目录，与train、val中的第一列进行拼接。

import os
from onekey_algo.classification.run_classification import main as clf_main
from collections import namedtuple

# 设置参数
root = r'E:\20240601-BiQiu\Pathomics'
# for model_name in ['densenet121']:
for model_name in ['CrossFormer']:
    train_f = os.path.join(root, 'split_info', 'train-RND-7.txt')
    val_f = os.path.join(root, 'split_info', 'val-RND-7.txt')
    labels_f = os.path.join(root, 'split_info', 'labels.txt')
    data_pattern = r''

    params = dict(train=train_f,
                  valid=val_f,
                  labels_file=labels_f,
                  data_pattern=data_pattern,
                  j=6,
                  max2use=None,
                  val_max2use=None,
                  batch_balance=False,
                  normalize_method='imagenet',
                  model_name=model_name,
                  gpus=[0],
                  batch_size=64,
                  epochs=4,
                  init_lr=0.001,
                  optimizer='sgd',
                  retrain=r'',
                  model_root=os.path.join(root, 'models'),
                  iters_start=0,
                  iters_verbose=256,
                  save_per_epoch=True,
                  add_date=False,
                  pretrained=True)
    # 训练模型
    Args = namedtuple("Args", params)
    clf_main(Args(**params))
